data-module add roomdb package, refactor
